+++ packages/x-adapter-platform/src/endpoint-adapters/identifier-results.endpoint-adapter.ts
@@ -1,9 +1,10 @@
-import { endpointAdapterFactory } from '@empathyco/x-adapter';
+import { endpointAdapterFactory, interpolate } from '@empathyco/x-adapter';
 import { IdentifierResultsRequest, IdentifierResultsResponse } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
 import { identifierResultsRequestMapper } from '../mappers/requests/identifier-results-request.mapper';
 // eslint-disable-next-line max-len
 import { identifierResultsResponseMapper } from '../mappers/responses/identifier-results-response.mapper';
+import { getSearchServiceUrl } from './utils';
 
 /**
  * Default adapter for the identifier results endpoint.
@@ -14,8 +15,8 @@ export const identifierResultsEndpointAdapter = endpointAdapterFactory<
   IdentifierResultsRequest,
   IdentifierResultsResponse
 >({
-  endpoint:
-    'https://api.{extraParams.env(.)}empathy.co/search/v1/query/{extraParams.instance}/skusearch',
+  endpoint: from =>
+    interpolate(`${getSearchServiceUrl(from)}/query/{extraParams.instance}/skusearch`, from),
   requestMapper: identifierResultsRequestMapper,
   responseMapper: identifierResultsResponseMapper,
   defaultRequestOptions: {