+++ packages/x-adapter-platform/src/endpoint-adapters/recommendations.endpoint-adapter.ts
@@ -1,8 +1,9 @@
-import { endpointAdapterFactory } from '@empathyco/x-adapter';
+import { endpointAdapterFactory, interpolate } from '@empathyco/x-adapter';
 import { RecommendationsRequest, RecommendationsResponse } from '@empathyco/x-types';
 import { recommendationsRequestMapper } from '../mappers/requests/recommendations-request.mapper';
 // eslint-disable-next-line max-len
 import { recommendationsResponseMapper } from '../mappers/responses/recommendations-response.mapper';
+import { getSearchServiceUrl } from './utils';
 
 /**
  * Default adapter for the recommendations' endpoint.
@@ -13,8 +14,8 @@ export const recommendationsEndpointAdapter = endpointAdapterFactory<
   RecommendationsRequest,
   RecommendationsResponse
 >({
-  endpoint:
-    'https://api.{extraParams.env(.)}empathy.co/search/v1/query/{extraParams.instance}/topclicked',
+  endpoint: from =>
+    interpolate(`${getSearchServiceUrl(from)}/query/{extraParams.instance}/topclicked`, from),
   requestMapper: recommendationsRequestMapper,
   responseMapper: recommendationsResponseMapper,
   defaultRequestOptions: {